restructure extensions to unify location
